Apply incoming transactions in parallel
Introduce a sequencer, a structure which demultiplexes subscribed
appliers into one sequential stream. Each next transaction is allowed to
be started if only the previous only was sent to wal. It helps to batch
replication.
Also a special txn_prepare method is introduced because of applier needs
to manage the exact order in which transaction are going to be written
despite of yields at engine_prepare phase.

Needed for: #1254