box: local sql_flags for parser and vdbe
The sql_flags is a parser parameter that describes how to parse
the SQL request, determines general behaviour: like whether
foreign keys are handled as deferred or not etc. But now this
information is taken from the global user session object.

When we need to run the parser with some other parameters, it is
necessary to change global session object, which may lead to
unpredictable consequences in general case.
Introduced a new parser and vdbe field sql_flags which is
responsible for SQL parsing results.

Needed for #3691